+++ src/components/Checkbox/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./Checkbox";
+export type { CheckboxConfig } from "./Checkbox";