+++ src/components/Typography/index.ts
@@ -1,1 +1,6 @@
-export { Heading, Text } from "./Typography";
+export { default as Heading } from "./Heading";
+export type { HeadingTag } from "./Heading";
+export { default as Body } from "./Body";
+export { default as Span } from "./Span";
+export { default as P } from "./P";
+export { default as Hr } from "./Hr";