+++ src/styles/theme/index.ts
@@ -1,0 +1,7 @@
+export { default } from "./default.theme";
+export type {
+  OakTheme,
+  OakColorName,
+  OakFontName,
+  PropsWithTheme,
+} from "./types";