+++ src/styles/utils/color.ts
@@ -1,7 +1,7 @@
 import { css } from "styled-components";
 
 import { OakColorName } from "../theme";
-import getColor from "../themeHelpers/getColor";
+import getColorByName from "../themeHelpers/getColorByName";
 
 import responsive, { ResponsiveValues } from "./responsive";
 
@@ -12,7 +12,7 @@ const color = css<ColorProps>`
   ${responsive<ColorProps, OakColorName>(
     "color",
     (props) => props.color,
-    getColor
+    getColorByName
   )}
 `;